internal: refactor a lot of stuff
More data is passed using contexts and so utility functions have
been added to do this in a type safe way. Stuff has been moved into
new packages to make it nicer.

Widgets have been into it's own package for good measure, and
Application is not a widget anymore. The Widget interface has been
simplified for mpdlrc's uses and Update now controls drawing.

This is all in preparation for adding widget.Status